crd-xray/requirements.txt
kopf>=1.35.0
kubernetes>=28.1.0
qdrant-client>=1.7.0
langchain-anthropic>=0.1.1
langgraph>=0.0.15
python-dotenv>=1.0.0
sentence-transformers>=2.2.2 